Added more cameras for further zoom outs. Fixed race conditions. ArrayPools now used to replace large array allocations every call. Added distance checks and agent reachable bool. These are now used to reduce the amount of GetPathCalls - especially noticeable when the player is standing or not moving much. Path queries can also be skipped couple times to reduce the MapGetPath() calls - especially with farther away agents.
Overall this slightly reduces the performance in general due to more data being copied around, but increases performance quite a bit in many circumstances and improves the stability.
These are just some ideas which anyone can implement on their own, remove, ... Not all of this is needed - other than the really necessary thread race conditions fixes.